requirements.txt
# 期货CTA策略依赖
# ============================

# 数据源
akshare>=1.10.0

# 数据处理
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.10.0

# 配置
pyyaml>=6.0

# Phase 3 盘中实时监控（可选，需 config.yaml 中 tqsdk 账户）
tqsdk>=3.5.0

# 自测
pytest>=7.0.0
